JDP191001-5 - review

+++ src/main/java/com/kodilla/ecommercee/UserDto.java
@@ -1,19 +1,19 @@
 package com.kodilla.ecommercee;
 
 public class UserDto {
-    private Long id;
+    private Integer id;
     private String username;
     private Integer status;
     private Long userKey;
 
-    public UserDto(Long id, String username, Integer status, Long userKey) {
+    public UserDto(Integer id, String username, Integer status, Long userKey) {
         this.id = id;
         this.username = username;
         this.status = status;
         this.userKey = userKey;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
